缓存过期时间字段改为long

+++ yunyue-core/yunyue-core-common/src/main/java/tech/yunyue/core/properties/Cache.java
@@ -34,7 +34,7 @@ public class Cache {
         /**
          * 缓存过期时间
          **/
-        private int secondsToExpire = 100;
+        private long secondsToExpire = 100;
 
         /*** 使用描述创建缓存 ***/
         private String spec;
